test(tunnel): move relay tests next to the code they cover
tunnel_relay_test.go collected tests for three different files: dialDirect
and validateTargetAddr live in tunnel_daemon.go, closeWriteSide lives in
tunnel_client.go. With the relay lifecycle now split across
tunnel_relay_unix.go and tunnel_relay_windows.go, the name promised platform
seam coverage it did not have.

Each test now sits beside its target: TestDialDirect and
TestValidateTargetAddr in tunnel_daemon_test.go, TestCloseWriteSide in
tunnel_client_test.go. The acceptOne helper moves to tunnel_daemon_test.go
and stays a single copy shared within package runner. No case was dropped and
no assertion changed.
