+++ pkg/runner/tunnel_client_test.go
@@ -3,8 +3,53 @@ package runner
 import (
 	"context"
 	"encoding/json"
+	"io"
+	"net"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+func TestCloseWriteSide(t *testing.T) {
+	t.Run("tcp conn delivers EOF but keeps read side open", func(t *testing.T) {
+		ln, err := net.Listen("tcp", "127.0.0.1:0")
+		require.NoError(t, err)
+		defer func() { _ = ln.Close() }()
+		serverCh := acceptOne(t, ln)
+
+		client, err := net.Dial("tcp", ln.Addr().String())
+		require.NoError(t, err)
+		defer func() { _ = client.Close() }()
+
+		server := <-serverCh
+		require.NotNil(t, server)
+		defer func() { _ = server.Close() }()
+
+		closeWriteSide(client)
+
+		// Server sees EOF on read...
+		buf := make([]byte, 1)
+		_, err = server.Read(buf)
+		assert.Equal(t, io.EOF, err)
+
+		// ...but the reverse direction still works.
+		_, err = server.Write([]byte("x"))
+		require.NoError(t, err)
+		n, err := client.Read(buf)
+		require.NoError(t, err)
+		assert.Equal(t, 1, n)
+		assert.Equal(t, byte('x'), buf[0])
+	})
+
+	t.Run("conn without CloseWrite is a safe no-op", func(t *testing.T) {
+		a, b := net.Pipe()
+		defer func() { _ = a.Close() }()
+		defer func() { _ = b.Close() }()
+
+		assert.NotPanics(t, func() { closeWriteSide(a) })
+	})
+}
 
 func TestIsValidSessionID(t *testing.T) {
 	tests := []struct {
